add ProfileFormcontent
