+++ components/organisms/ProfileForm/ProfileForm.js
@@ -1,69 +1,10 @@
 import React from 'react';
-import * as Yup from 'yup';
-import styled from 'styled-components';
 import { useUpdateUser } from 'lib/apollo/hooks/actions';
+import ProfileFormContent from './ProfileFormContent';
 import ErrorDecorator from 'decorators/ErrorDecorator';
-import Form from '../../molecules/Form';
 
-const StyledTitle = styled.h3`
-  max-width: 40rem;
-  margin: 0 auto 1rem;
-  line-height: 1.25;
-  letter-spacing: -0.035em;
-`;
-
-const ProfileForm = ({ profile: { email, firstName, lastName } }) => {
+const ProfileForm = ({ profile }) => {
   const [updateUser] = useUpdateUser();
-
-  const fields = [
-    {
-      type: 'text',
-      name: 'firstName',
-      title: 'First Name',
-      placeholder: 'First Name',
-      initialValue: firstName || '',
-      validationSchema: Yup.string(),
-    },
-    {
-      type: 'text',
-      name: 'lastName',
-      title: 'Last Name',
-      placeholder: 'Last Name',
-      initialValue: lastName || '',
-      validationSchema: Yup.string(),
-    },
-    {
-      type: 'email',
-      name: 'email',
-      title: 'Email',
-      placeholder: 'Email',
-      initialValue: email || '',
-      validationSchema: Yup.string()
-        .email('The email must be valid!!')
-        .required('This field is required'),
-    },
-    {
-      type: 'password',
-      name: 'password',
-      title: 'New Password',
-      placeholder: 'New Password',
-      initialValue: '',
-      validationSchema: Yup.string(),
-    },
-    {
-      type: 'currentPassword',
-      name: 'currentPassword',
-      title: 'Confirm Password',
-      placeholder: 'Confirm Password',
-      initialValue: '',
-      validationSchema: Yup.string(),
-    },
-    {
-      type: 'submit',
-      name: 'Update',
-      initialValue: 'Update',
-    },
-  ];
 
   const onSubmit = async (values, { setSubmitting, setStatus }) => {
     try {
@@ -78,16 +19,6 @@ const ProfileForm = ({ profile: { email, firstName, lastName } }) => {
     }
   };
 
-  const form = {
-    fields,
-    submit: onSubmit,
-  };
-
-  return (
-    <>
-      <StyledTitle>Profile</StyledTitle>
-      <Form form={form} />
-    </>
-  );
+  return <ProfileFormContent profile={profile} onSubmit={onSubmit} />;
 };
 export default ProfileForm;
